// packages/receipt/src/generated/types.ts
export interface CryptoType {
  /** Text content of the element. */
  value: string;
  cryptoType: string;
}
// ...
export interface SignedDocumentEnvelopeType {
  /**
   * Interface version of the envelope's content.
   *
   * A boríték tartalmának interfész verziója.
   */
  envelopeVersion?: string;
  /**
   * Document data in BASE64 encoded form (see: CoreDocument)
   *
   * Bizonylat adatok BASE64-ben kódolt tartalma (lásd: CoreDocument)
   */
  envelopeData: string;
  /**
   * Customer data in BASE64 encoded form (see: CustomerDocument)
   *
   * Vevői adatok BASE64-ben kódolt tartalma (lásd: CustomerDocument)
   */
  customerEnvelopeData: string;
  /**
   * Envelope file (envelopeData and customerEnvelopeData) hash value
   *
   * Boriték állomány (envelopeData és customerEnvelopeData) hash lenyomata
   */
  envelopeHash: CryptoType;
  /**
   * Envelope signature (envelopeData and customerEnvelopeData) in BASE64 encoded
   * form
   *
   * Boriték adatok (envelopeData és customerEnvelopeData) BASE64-ben kódolt
   * aláirása
   */
  envelopeSignature: string;
}

/**
 * Signed envelope type
 *
 * Alap boriték tipus
 */
export interface SignedReportEnvelopeType {
  /**
   * Interface version of the envelope's content.
   *
   * A boríték tartalmának interfész verziója.
   */
  envelopeVersion?: string;
  /**
   * Report data in BASE64 encoded form (see: CoreReport)
   *
   * Riport adatok BASE64-ben kódolt tartalma (lásd: CoreReport)
   */
  envelopeData: string;
  /**
   * Envelope data in BASE64 encoded form (see: CustomerReport)
   *
   * Boriték adatok BASE64-ben kódolt tartalma (lásd: CustomerReport)
   */
  customerEnvelopeData?: string;
  /**
   * Envelope file (envelopeData and customerEnvelopeData) hash value
   *
   * Boriték állomány (envelopeData és customerEnvelopeData) hash lenyomata
   */
  envelopeHash: CryptoType;
  /**
   * Envelope signature (envelopeData and customerEnvelopeData) in BASE64 encoded
   * form
   *
   * Boriték adatok (envelopeData és customerEnvelopeData) BASE64-ben kódolt
   * aláirása
   */
  envelopeSignature: string;
}